Add some mediocre tests (#23)
* Add circle badge

* Extract tests

* Style

* Fix test import on remote

* Don't copy code because we lose cov report

* Coverage steroids

* Style

* Add ddl factory test

+++ extract/parsers/util.py
@@ -4,10 +4,10 @@ from pathlib import Path
 import zstd
 
 
-OUTPUT_PATH = Path("/parsed")
+OUTPUT_PATH = Path("parsed")
 
 
-def compress(file: Path, output_dir: Path) -> None:
+def compress(file: Path, output_dir: Path, remove_original=True) -> None:
     """Replaces the original file with a compressed version"""
     logging.info("Compressing {}".format(file))
     compressed_file = output_dir.joinpath(file.stem).with_suffix(file.suffix + ".zst")
@@ -15,4 +15,5 @@ def compress(file: Path, output_dir: Path) -> None:
     with open(file, 'rb') as ifh, open(compressed_file, 'wb') as ofh:
         compression_result = cctx.copy_stream(ifh, ofh)
         print("{} size (uncompressed,compressed): {}".format(file, compression_result))
-    return file.unlink()
+    if remove_original:
+        return file.unlink()
